better reference tracking in new_from_memory
We were using "in" to check for repeats of memoryview objects, but "in"
uses "==", and "==" on memoryview compares hashes. This means that
different pointers are equal if they point to equal bytes.

This patch swaps "in" for a loop testing equality with "is" (pointer
equality). It should speed up pyvips operations on large memoryviews
too, since hashes will no longer be calculated.

Thanks aspadm

See https://github.com/libvips/pyvips/issues/212